Fix nested Permission request
If a second Permission request is added inside the Runnable of a first request,
its request block then gets evalauated immediately using the grant results of
the first block, which means that the fallback runnable of the second block
might get executed erroneously.

As a workaround/fix, we therefore spin the event loop once.