Use JmDNS to broadcast our hostname via mDNS
It turns out that registering an mDNS service via the NsdManager only does just
that, but doesn't simultaneously register the service name as an additional
hostname for DNS resolution purposes.

It seems there aren't any further APIs for that purpose, so the only solution is
running our own custom mDNS responder. Luckily mDNS uses a port >= 1024 and due
to multicast both the system DNS-SD and our own implementation running in
parallel on the same port isn't a problem either.

Therefore we use JmDNS to be able to configure a custom hostname to advertise
via mDNS, making the text in our UI actually true.