Update JFileServer dependency to fix Android 6 crash
The core lib desugaring doesn't completely cover the change in Java 8 of
(Concurrent)Hashmap.keySet() now returning a KeySetView instead of simply a Set.
AFAIK API24+ is fine, but API23 (Android 6) unfortunately isn't.

JFileServer accesses a HashMap's keySet() in a few places, but luckily the only
known crash so far in practical usage is when shutting down the server with
files still open on the file share.

Luckily in this case the old array-based OpenFileMap implementation was still
around in the codebase, so we've wired it up to make its usage configurable.